chore: Update routing to load EnterpriseSelectModule instead of OptionListModule
