+++ apps/picsa-tools/option-tool/src/app/pages/enterprise-select/enterprise-select.component.html
@@ -3,4 +3,5 @@
   <div *ngFor="let option of options" class="options-container">
     <a [routerLink]="[option.id]">{{ option.label }}</a>
   </div>
+  <mat-icon svgIcon="picsa_enterprise:crop"></mat-icon>
 </div>
