+++ apps/picsa-tools/option-tool/src/app/pages/enterprise-select/enterprise-select.component.scss
