+++ apps/picsa-tools/option-tool/src/app/pages/enterprise-select/enterprise-select.module.ts
@@ -1,0 +1,20 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { PicsaTranslateModule } from '@picsa/shared/modules';
+
+import { OptionToolComponentsModule } from '../../components/components.module';
+import { EnterpriseSelectComponent } from './enterprise-select.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: EnterpriseSelectComponent,
+  },
+];
+
+@NgModule({
+  declarations: [EnterpriseSelectComponent],
+  imports: [CommonModule, RouterModule.forChild(routes), OptionToolComponentsModule, PicsaTranslateModule],
+})
+export class EnterpriseSelectModule {}
